requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-Migrate==4.0.5
python-dotenv==1.0.0
Werkzeug==2.3.7
SQLAlchemy==2.0.21
PyJWT==2.8.0
bcrypt==4.0.1
python-dateutil==2.8.2
gunicorn==21.2.0
psycopg2-binary==2.9.9 